requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
dnspython==2.2.1
django==4.1
django-environ==0.9.0
djangorestframework==3.13.1
djongo==1.3.6
pymongo==3.12.3
pytz==2022.2.1
requests-html==0.10.0
sqlparse==0.2.4
